Support HUGE files in spcomp (bug 5959, r=dvander)
Summary: When there are more than 32768 SLOC in a file, spcomp underflows when storing debug info. These are safe to change as the file-backed struct uses uint32_t already.

Test Plan: Compiled the test plugin from bug 5959. I didn't load it, but it compiled without error and the SMX file is valid.

Reviewers: BAILOPAN

Differential Revision: https://phabricator.alliedmods.net/D2

--HG--
extra : rebase_source : fa907333b8c0dcab6e78c69ad4a92cdfb048083f
extra : amend_source : c7d2ec7d8b880459a57e4b2df5c1e8bbab5585bb

+++ sourcepawn/compiler/amxdbg.h
@@ -69,8 +69,8 @@ typedef struct tagAMX_DBG_HDR {
   char    amx_version;          /* required version of the AMX */
   int16_t flags         PACKED; /* currently unused */
   int16_t files         PACKED; /* number of entries in the "file" table */
-  int16_t lines         PACKED; /* number of entries in the "line" table */
-  int16_t symbols       PACKED; /* number of entries in the "symbol" table */
+  int32_t lines         PACKED; /* number of entries in the "line" table */
+  int32_t symbols       PACKED; /* number of entries in the "symbol" table */
   int16_t tags          PACKED; /* number of entries in the "tag" table */
   int16_t automatons    PACKED; /* number of entries in the "automaton" table */
   int16_t states        PACKED; /* number of entries in the "state" table */
